Added a container class for convenience
The .dcg-container class creates an element in the middle of the screen horizontally by combining .dcg-s10 .dcg-m10 .dcg-l8 and .dcg-center together into one class. The styles associated with these classes will be applied automatically by just using the one class.

The s10 m10 and l8 classes are responsive classes that define the width of columns on different sized screens, and the center class centers the element and its contents. You may want to apply .dcg-left-align to make text on the left side again. (Or .dcg-right-align if you prefer the right)