Started Documentation and Working Demos and Examples
Decided to start working on the documentation for the project, which contains working demos and copy-and-paste-able code examples.

All the documentation will be housed under the one file using switch statements to dynamically change content that will be hard coded. May change this in the future, but for now it works.